+++ Assets/Scripts/Network/PacketArgs/ProfileResponse.cs
@@ -4,7 +4,7 @@ namespace Network.PacketArgs
     public class ProfileResponse
     {
         public int AccountId { get; set; }
-        public int Gold { get; set; }
-        public int Diamonds { get; set; }
+        public long Gold { get; set; }
+        public long Diamonds { get; set; }
     }
 }
